requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
lighter-sdk>=1.0.4
